+++ src/backend/plugins/video/tests/test_api.py
@@ -1,9 +1,12 @@
 """Tests for the video API endpoints."""
 
+import arrow
 import pytest
+from warren_video.api import VideoDayViews, VideoViews
 
 from warren.conf import settings
 from warren.factories.video import VideoPlayedFactory
+from warren.filters import DatetimeRange
 
 
 @pytest.mark.anyio
@@ -43,11 +46,20 @@ async def test_views_valid_video_id_path_but_no_matching_video(
 ):
     """Test the video views endpoint with a valid "video_id" but no results."""
 
-    no_statements_response = {"daily_views": [], "total": 0}
+    default_range = DatetimeRange()
+    no_statements_response = VideoViews(
+        total=0,
+        daily_views=[
+            VideoDayViews(day=day.format("YYYY-MM-DD"))
+            for day in arrow.Arrow.range(
+                "day", default_range.since, default_range.until
+            )
+        ],
+    )
 
     response = await http_client.get(f"/api/v1/video/{video_id}/views")
     assert response.status_code == 200
-    assert response.json() == no_statements_response
+    assert VideoViews.parse_obj(response.json()) == no_statements_response
 
 
 @pytest.mark.anyio
@@ -67,6 +79,7 @@ async def test_views_backend_query(es_client, http_client):
                             }
                         }
                     },
+                    {"timestamp": "2023-02-03T15:34:00.000-01:00"},
                 ]
             ),
             VideoPlayedFactory.build(
@@ -79,6 +92,7 @@ async def test_views_backend_query(es_client, http_client):
                             }
                         }
                     },
+                    {"timestamp": "2023-02-07T19:34:00.000-01:00"},
                 ]
             ),
             VideoPlayedFactory.build(
@@ -91,6 +105,7 @@ async def test_views_backend_query(es_client, http_client):
                             }
                         }
                     },
+                    {"timestamp": "2023-02-07T15:28:00.000-01:00"},
                 ]
             ),
             VideoPlayedFactory.build(
@@ -103,9 +118,21 @@ async def test_views_backend_query(es_client, http_client):
     await es_client.indices.refresh(index=settings.ES_INDEX)
 
     response = await http_client.get(
-        "/api/v1/video/uuid://ba4252ce-d042-43b0-92e8-f033f45612ee/views"
+        "/api/v1/video/uuid://ba4252ce-d042-43b0-92e8-f033f45612ee/views",
+        params={"since": "2023-02-01", "until": "2023-02-10"},
     )
     assert response.status_code == 200
-    daily_views = response.json()
-    assert daily_views.get("total") == 2
-    assert daily_views.get("daily_views") == [{"day": "2021-12-01", "views": 2}]
+    video_views = VideoViews.parse_obj(response.json())
+    assert video_views.total == 2
+    assert video_views.daily_views == [
+        VideoDayViews(day="2023-02-01", views=0),
+        VideoDayViews(day="2023-02-02", views=0),
+        VideoDayViews(day="2023-02-03", views=1),
+        VideoDayViews(day="2023-02-04", views=0),
+        VideoDayViews(day="2023-02-05", views=0),
+        VideoDayViews(day="2023-02-06", views=0),
+        VideoDayViews(day="2023-02-07", views=1),
+        VideoDayViews(day="2023-02-08", views=0),
+        VideoDayViews(day="2023-02-09", views=0),
+        VideoDayViews(day="2023-02-10", views=0),
+    ]